fix(aiand): preserve curated modalities, skip thin standalone creates, move tests to sync.test.ts
